get_container_dir moved to utils
