+++ src/neptune/internal/operation_processors/sync_operation_processor.py
@@ -25,11 +25,11 @@ from typing import (
 from neptune.constants import SYNC_DIRECTORY
 from neptune.core.components.metadata_file import MetadataFile
 from neptune.internal.operation_processors.operation_processor import OperationProcessor
-from neptune.internal.operation_processors.operation_storage import (
-    OperationStorage,
+from neptune.internal.operation_processors.operation_storage import OperationStorage
+from neptune.internal.operation_processors.utils import (
+    common_metadata,
     get_container_dir,
 )
-from neptune.internal.operation_processors.utils import common_metadata
 from neptune.internal.utils.disk_utilization import ensure_disk_not_overutilize
 from neptune.internal.utils.files import should_clean_internal_data
