+++ tests/unit/neptune/new/client/test_run.py
@@ -111,7 +111,7 @@ class TestClientRun(AbstractExperimentTestMixin, unittest.TestCase):
     )
     @unittest.skipIf(IS_WINDOWS, "Linux/Mac test")
     @patch("neptune.internal.operation_processors.operation_storage.os.listdir", new=lambda path: [])
-    @patch("neptune.internal.metadata_file.open", mock_open())
+    @patch("neptune.core.components.metadata_file.open", mock_open())
     def test_entrypoint(self):
         with init_run(mode="debug") as exp:
             self.assertEqual(exp["source_code/entrypoint"].fetch(), "main.py")
@@ -201,7 +201,7 @@ class TestClientRun(AbstractExperimentTestMixin, unittest.TestCase):
         new=lambda path: os.path.normpath(os.path.join("/home/user/main_dir", path)),
     )
     @patch("neptune.internal.operation_processors.operation_storage.os.listdir", new=lambda path: [])
-    @patch("neptune.internal.metadata_file.open", mock_open())
+    @patch("neptune.core.components.metadata_file.open", mock_open())
     def test_entrypoint_without_common_root(self):
         with init_run(mode="debug", source_files=["../*"]) as exp:
             self.assertEqual(exp["source_code/entrypoint"].fetch(), "/home/user/main_dir/main.py")
